move models to model
