+++ article/app/model/LiveBlogPage.scala
@@ -1,5 +1,3 @@
-package models
-
-import model.{Article, RelatedContent}
+package model
 
 case class LiveBlogPage(article: Article, currentPage: LiveBlogCurrentPage, related: RelatedContent) extends PageWithStoryPackage
