+++ article/app/model/ParseBlockId.scala
@@ -1,4 +1,4 @@
-package models
+package model
 
 import scala.util.parsing.combinator.RegexParsers
