+++ article/app/views/articleEmail.scala.html
@@ -1,3 +1,4 @@
+@import model.PageWithStoryPackage
 
 @(page: PageWithStoryPackage)(implicit request: RequestHeader, context: model.ApplicationContext)
