+++ article/app/views/fragments/emailArticleBody.scala.html
@@ -1,3 +1,5 @@
+@import model.PageWithStoryPackage
+
 @(page: PageWithStoryPackage)(implicit request: RequestHeader)
 
 @import views.support.{EmailImage, EmailVideoImage}
