+++ article/app/views/fragments/liveBlogNavigation.scala.html
@@ -1,4 +1,5 @@
 @import conf.Configuration
+@import model.LiveBlogCurrentPage
 
 @(id: String, currentPage: LiveBlogCurrentPage, amp: Boolean = false)
