+++ article/app/views/liveBlog.scala.html
@@ -1,4 +1,7 @@
 
+@import model.LiveBlogPage
+@import model.LiveBlogHelpers
+
 @(model: LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
 
 @import common.LinkTo
