+++ article/app/views/liveBlogAMP.scala.html
@@ -1,4 +1,6 @@
-@(model: models.LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
+@import model.LiveBlogPage
+
+@(model: LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
 
 @mainAMP(model, model.related, model.article.content){
     @views.html.liveblog.liveBlogBody(model, amp = true)
