+++ article/app/views/liveblog/keyEvents.scala.html
@@ -1,4 +1,4 @@
-@import models.KeyEventData
+@import model.KeyEventData
 @(id: String, keyEvents: Seq[KeyEventData])
 
 @keyEvents.map { keyEvent =>
