+++ article/app/views/liveblog/liveBlogBody.scala.html
@@ -1,6 +1,6 @@
 
 @import model.Badges.badgeFor
-@import models.{KeyEventData, LiveBlogPage}
+@import model.{KeyEventData, LiveBlogPage}
 
 @(model: LiveBlogPage, amp: Boolean = false)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
