+++ article/app/views/liveblog/liveBlogBodyContent.scala.html
@@ -1,4 +1,6 @@
 
+@import model.{LiveBlogPage, LatestBlock}
+
 @(model: LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
 
 @import common.Edition
