+++ article/app/views/liveblog/liveBlogBodyContentAMP.scala.html
@@ -1,4 +1,5 @@
-@import model.Article
+@import model.{Article, LiveBlogPage, LatestBlock}
+
 @(model: LiveBlogPage)(implicit request: RequestHeader, context: _root_.model.ApplicationContext)
 
 @import common.Edition
